# Conflicts:
#	build/linux/debian/DEBIAN/control
#	build/linux/linux-arm
#	build/linux/linux-x64
#	build/linux/papdy.deb
#	build/mac/mac-arm
#	build/mac/mac-x64
#	build/papdy.phar
#	build/windows/windows-x64.exe
#	config/app.php

+++ app/Models/Process/Deploy.php
@@ -41,16 +41,8 @@ class Deploy extends Base
     /**
      * @param array<int, string> $servers
      */
-    public function execute(OutputInterface $output, array $servers, string $name): void
+    public function execute(OutputInterface $output, array $servers, string $name, ?string $localBuildNameFile): void
     {
-        $buildServerBuildNameFile = $this->runScript(
-            $output,
-            'deploy/build.sh',
-            [],
-            ['build:single'],
-            ['name' => $name]
-        );
-
         $localBuildPath = storage_path('build');
 
         if (str_contains($localBuildPath, 'phar://')) {
@@ -69,15 +61,29 @@ class Deploy extends Base
             $this->files->createDirectory($localBuildPath, 0755);
         }
 
-        $localBuildNameFile = sprintf('%s/%s', $localBuildPath, basename($buildServerBuildNameFile));
+        if ($this->variables->isEmpty($localBuildNameFile)) {
+            $buildServerBuildNameFile = $this->runScript(
+                $output,
+                'deploy/build.sh',
+                [],
+                ['build:single'],
+                ['name' => $name]
+            );
 
-        $this->download($output, $buildServerBuildNameFile, $localBuildNameFile, ['build:single']);
+            $localBuildNameFile = sprintf('%s/%s', $localBuildPath, basename($buildServerBuildNameFile));
+
+            $this->download($output, $buildServerBuildNameFile, $localBuildNameFile, ['build:single']);
+
+            $removeLocalFile = true;
+        } else {
+            $removeLocalFile = false;
+        }
 
         $deployServerBuildNameFile = sprintf(
             '%s%s%s',
             sys_get_temp_dir(),
             DIRECTORY_SEPARATOR,
-            basename($buildServerBuildNameFile),
+            basename($localBuildNameFile),
         );
 
         $deployServer = sprintf('deploy:%s', count($servers) > 0 ? implode(',', $servers) : 'all');
@@ -100,8 +106,10 @@ class Deploy extends Base
 
         $this->delete($output, $deployServerBuildNameFile, [$deployServer]);
 
-        $output->writeln(sprintf('Deleting file at: %s', $localBuildNameFile));
+        if ($removeLocalFile) {
+            $output->writeln(sprintf('Deleting file at: %s', $localBuildNameFile));
 
-        unlink($localBuildNameFile);
+            unlink($localBuildNameFile);
+        }
     }
 }
